+++ include/mn_auto_version.hpp
@@ -4,23 +4,23 @@
 namespace mn{
 	
 	//Software Status
-	static const char MN_VERSION_STATUS[] =  "Release";
-	static const char MN_VERSION_STATUS_SHORT[] =  "r";
+	#define MN_VERSION_STATUS  "Beta"
+	#define MN_VERSION_STATUS_SHORT  "b"
 	
 	//Standard Version Type
-	static const long MN_VERSION_MAJOR  = 2;
-	static const long MN_VERSION_MINOR  = 28;
-	static const long MN_VERSION_BUILD  = 3098;
-	static const long MN_VERSION_REVISION  = 0;
+	#define MN_VERSION_MAJOR  2
+	#define MN_VERSION_MINOR  29
+	#define MN_VERSION_BUILD  8906
+	#define MN_VERSION_REVISION  7
 	
 	//Miscellaneous Version Types
-	static const long MN_VERSION_BUILDS_COUNT  = 30;
-	#define MN_VERSION_RC_FILEVERSION 2,28,3098,0
-	#define MN_VERSION_RC_FILEVERSION_STRING "2, 28, 3098, 0\0"
-	static const char MN_VERSION_FULLVERSION_STRING [] = "2.28.3098.0";
+	#define MN_VERSION_BUILDS_COUNT  129
+	#define MN_VERSION_RC_FILEVERSION 2,29,8906,7
+	#define MN_VERSION_RC_FILEVERSION_STRING "2, 29, 8906, 7\0"
+	#define MN_VERSION_FULLVERSION_STRING  "2.29.8906.7"
 	
 	//These values are to keep track of your versioning state, don't modify them.
-	static const long MN_VERSION_BUILD_HISTORY  = 1;
+	#define MN_VERSION_BUILD_HISTORY  1
 	
 
 }